+++ src/paco/models/loader.py
@@ -45,6 +45,7 @@ from paco.models.applications import Application, PinpointApplication, ResourceG
     ElasticsearchDomain, ElasticsearchCluster, EBSOptions, ESAdvancedOptions, \
     ECSContainerDefinition, ECSContainerDefinitions, ECSTaskDefinitions, ECSTaskDefinition, \
     ECSLoadBalancer, ECSServicesContainer, ECSService, ECSCluster, ECSServices, PortMapping, ECSMountPoint, \
+    ECSTargetTrackingScalingPolicies, ECSTargetTrackingScalingPolicy, \
     ECSVolumesFrom, ECSVolume, ECSLogging, ECRRepository, ECSTaskDefinitionSecret, ECSContainerDependency, \
     DockerLabels, ECSHostEntry, ECSHealthCheck, ECSUlimit, ECSCapacityProvider, ServicesMonitorConfig,\
     PinpointSMSChannel, PinpointEmailChannel
@@ -189,6 +190,7 @@ SUB_TYPES_CLASS_MAP = {
     },
     ECSService: {
         'load_balancers': ('obj_list', ECSLoadBalancer),
+        'target_tracking_scaling_policies': ('container', (ECSTargetTrackingScalingPolicies, ECSTargetTrackingScalingPolicy)),
         'monitoring': ('direct_obj', MonitorConfig),
     },
     ECSTaskDefinition: {